+++ src/Database/Adapter/PdoAdapter.php
@@ -35,12 +35,11 @@ abstract class PdoAdapter implements AdapterInterface
         } else {
             $res = $this->pdo->query($sql);
         }
-        
-        if ($res !== false) {
-            return $res;
+
+        if ($res === false) {
+            $this->throwError($sql);
         }
-        
-        $this->throwError($sql);
+        return $res;
     }
 
     /**
@@ -49,14 +48,8 @@ abstract class PdoAdapter implements AdapterInterface
     public function insert($table, array $data)
     {
         $statement = $this->buildInsertQuery($table, $data);
-        if (!$statement) {
-            $this->throwError($statement);
-        }
-        $res = $this->execute($statement);
-        if ($res !== false) {
-            return $this->pdo->lastInsertId();
-        }
-        $this->throwError($statement->queryString);
+        $this->execute($statement);
+        return $this->pdo->lastInsertId();
     }
     
     /**
@@ -66,6 +59,9 @@ abstract class PdoAdapter implements AdapterInterface
     {
         $query = sprintf('INSERT INTO %s %s VALUES %s;', $this->queryBuilder->escapeString(addslashes($table)), $this->createKeys($data), $this->createValues($data));
         $statement = $this->pdo->prepare($query);
+        if (!$statement) {
+            $this->throwError($statement);
+        }
         foreach ($data as $key => $value) {
             if ($value instanceof DateTime) {
                 $value = $value->format('Y-m-d H:i:s');
@@ -81,9 +77,6 @@ abstract class PdoAdapter implements AdapterInterface
     public function update($table, array $data, array $conditions = [], $where = '')
     {
         $statement = $this->buildUpdateQuery($table, $data, $conditions, $where);
-        if (!$statement) {
-            $this->throwError($statement);
-        }
         return $this->execute($statement);
     }
     
@@ -98,6 +91,9 @@ abstract class PdoAdapter implements AdapterInterface
         }
         $query = sprintf('UPDATE %s SET %s%s;', $this->queryBuilder->escapeString(addslashes($table)), implode(', ', $values), $this->createWhere($conditions, $where));
         $statement = $this->pdo->prepare($query);
+        if (!$statement) {
+            $this->throwError($statement);
+        }
         foreach ($data as $key => $value) {
             if ($value instanceof DateTime) {
                 $value = $value->format('Y-m-d H:i:s');
@@ -116,9 +112,6 @@ abstract class PdoAdapter implements AdapterInterface
     public function delete($table, array $conditions = [], $where = '')
     {
         $statement = $this->buildDeleteQuery($table, $conditions, $where);
-        if (!$statement) {
-            $this->throwError($statement);
-        }
         return $this->execute($statement);
     }
     
@@ -129,6 +122,9 @@ abstract class PdoAdapter implements AdapterInterface
     {
         $query = sprintf('DELETE FROM %s%s;', $this->queryBuilder->escapeString(addslashes($table)), $this->createWhere($conditions, $where));
         $statement = $this->pdo->prepare($query);
+        if (!$statement) {
+            $this->throwError($statement);
+        }
         foreach ($conditions as $key => $condition) {
             $statement->bindValue('where_' . $key, $condition);
         }
@@ -151,8 +147,9 @@ abstract class PdoAdapter implements AdapterInterface
      */
     public function fetch($table, $fields = '*', array $conditions = [], array $orders = [], array $groups = [])
     {
-        $query = $this->buildFetchQuery($table, $fields, $conditions, 1, $orders, $groups);
-        return $this->execute($query)->fetch(PDO::FETCH_ASSOC);
+        $statement = $this->buildFetchQuery($table, $fields, $conditions, 1, $orders, $groups);
+        $this->execute($statement);
+        return $statement->fetch(PDO::FETCH_ASSOC);
     }
     
     /**
@@ -160,16 +157,24 @@ abstract class PdoAdapter implements AdapterInterface
      */
     public function fetchAll($table, $fields = '*', array $conditions = [], $limit = null, array $orders = [], array $groups = [])
     {
-        $query = $this->buildFetchQuery($table, $fields, $conditions, $limit, $orders, $groups);
-        return $this->execute($query)->fetchAll(PDO::FETCH_ASSOC);
+        $statement = $this->buildFetchQuery($table, $fields, $conditions, $limit, $orders, $groups);
+        $this->execute($statement);
+        return $statement->fetchAll(PDO::FETCH_ASSOC);
     }
 
     private function buildFetchQuery($table, $fields, array $conditions = [], $limit = null, array $orders = [], array $groups = [])
     {
-        $query = sprintf('SELECT %s FROM %s%s%s;', $fields, $this->queryBuilder->escapeString(addslashes($table)), $this->createWhere($conditions), $this->createLimit($limit));
-        return $query;
+        $query = sprintf('SELECT %s FROM %s%s%s%s%s;', $fields, $this->queryBuilder->escapeString(addslashes($table)), $this->createWhere($conditions), $this->createGroup($groups), $this->createOrder($orders), $this->createLimit($limit));
+        $statement = $this->pdo->prepare($query);
+        if (!$statement) {
+            $this->throwError($statement);
+        }
+        foreach ($conditions as $key => $condition) {
+            $statement->bindValue('where_' . $key, $condition);
+        }
+        return $statement;
     }
-    
+
     private function createKeys($data)
     {
         $keys = [];
@@ -178,7 +183,7 @@ abstract class PdoAdapter implements AdapterInterface
         }
         return '(' . implode(', ', $keys) . ')';
     }
-    
+
     private function createValues($data)
     {
         $values = [];
@@ -187,12 +192,12 @@ abstract class PdoAdapter implements AdapterInterface
         }
         return '(' . implode(', ', $values) . ')';
     }
-    
+
     private function createValue($key, $prefix = '')
     {
         return ':' . $prefix . $key;
     }
-    
+
     private function createWhere(array $conditions = [], $where = '')
     {
         if (empty($conditions) && $where == '') {
@@ -208,7 +213,7 @@ abstract class PdoAdapter implements AdapterInterface
         }
         return sprintf(' WHERE %s', implode(' AND ', $cond) . ($where ? ' AND ' . $where : ''));
     }
-    
+
     private function createLimit($limit = null)
     {
         if (!$limit) {
@@ -216,7 +221,32 @@ abstract class PdoAdapter implements AdapterInterface
         }
         return sprintf(' LIMIT %s', $limit);
     }
-    
+
+    private function createOrder(array $orders = [])
+    {
+        if (empty($orders)) {
+            return '';
+        }
+        $listOfOrders = [];
+        foreach ($orders as $column => $way) {
+            if (!in_array(strtoupper($way), ['ASC', 'DESC'])) {
+                $column = $way;
+                $way = 'ASC';
+            }
+            $way = strtoupper($way);
+            $listOfOrders[] = $column . ' ' . $way;
+        }
+        return sprintf(' ORDER BY %s', implode(', ', $listOfOrders));
+    }
+
+    private function createGroup(array $groups = [])
+    {
+        if (empty($groups)) {
+            return '';
+        }
+        return sprintf(' GROUP BY %s', implode(', ', $groups));
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -244,7 +274,7 @@ abstract class PdoAdapter implements AdapterInterface
     private function throwError($query)
     {
         $errorInfo = $this->pdo->errorInfo();
-        throw new DatabaseQueryExecuteException('SQLSTATE[' . $errorInfo[0] . ']: ' . $errorInfo[2] . '. Query ' . print_R($query, true) . ' fails', $errorInfo[1]);
+        throw new DatabaseQueryExecuteException('SQLSTATE[' . $errorInfo[0] . ']: ' . $errorInfo[2] . '.' . ($query ? ' Query ' . print_R($query, true) . ' fails' : ''), $errorInfo[1]);
     }
     
     /**