+++ tests/Database/QueryBuilder/PgsqlQueryBuilderTest.php
@@ -17,9 +17,9 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $table->addPrimary(true);
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'unsupported')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $this->setExpectedException('\Exception', 'Type "unsupported" is not allowed');
-        $queryCreator->createTable($table);
+        $queryBuilder->createTable($table);
     }
 
     public function testSimpleCreate()
@@ -28,12 +28,12 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $table->addPrimary(true);
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'string')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE SEQUENCE "simple_seq";',
             'CREATE TABLE "simple" ("id" int4 DEFAULT nextval(\'simple_seq\'::regclass) NOT NULL,"title" varchar(255) NOT NULL,CONSTRAINT "simple_pkey" PRIMARY KEY ("id"));'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testMoreColumns()
@@ -45,12 +45,12 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('total', 'integer', ['default' => 0])));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('bodytext', 'text')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE SEQUENCE "more_columns_seq";',
             'CREATE TABLE "more_columns" ("id" int4 DEFAULT nextval(\'more_columns_seq\'::regclass) NOT NULL,"title" varchar(255) NOT NULL,"alias" varchar(255) DEFAULT NULL,"total" int4 DEFAULT 0 NOT NULL,"bodytext" text NOT NULL,CONSTRAINT "more_columns_pkey" PRIMARY KEY ("id"));',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testNoPrimaryKey()
@@ -60,11 +60,11 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('total', 'integer', ['default' => 0])));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('is_deleted', 'boolean', ['default' => false])));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE TABLE "no_primary_key" ("title" varchar(255) DEFAULT NULL,"total" int4 DEFAULT 0 NOT NULL,"is_deleted" bool DEFAULT false NOT NULL);'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testOwnPrimaryKey()
@@ -73,11 +73,11 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $table->addPrimary(new Column('identifier', 'string', ['length' => 32]));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'string', ['default' => ''])));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE TABLE "own_primary_key" ("identifier" varchar(32) NOT NULL,"title" varchar(255) DEFAULT \'\' NOT NULL,CONSTRAINT "own_primary_key_pkey" PRIMARY KEY ("identifier"));'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testMoreOwnPrimaryKeys()
@@ -86,11 +86,11 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $table->addPrimary([new Column('identifier', 'string', ['length' => 32]), new Column('subidentifier', 'string', ['length' => 32])]);
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'string', ['default' => ''])));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE TABLE "more_own_primary_keys" ("identifier" varchar(32) NOT NULL,"subidentifier" varchar(32) NOT NULL,"title" varchar(255) DEFAULT \'\' NOT NULL,CONSTRAINT "more_own_primary_keys_pkey" PRIMARY KEY ("identifier","subidentifier"));'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testOneFieldAsPrimaryKey()
@@ -100,11 +100,11 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('identifier', 'string', ['length' => 32])));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'string', ['default' => ''])));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE TABLE "one_field_as_pk" ("identifier" varchar(32) NOT NULL,"title" varchar(255) DEFAULT \'\' NOT NULL,CONSTRAINT "one_field_as_pk_pkey" PRIMARY KEY ("identifier"));',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testMoreFieldsAsPrimaryKeys()
@@ -115,11 +115,11 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('subidentifier', 'string', ['length' => 32])));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'string', ['default' => ''])));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE TABLE "more_fields_as_pk" ("identifier" varchar(32) NOT NULL,"subidentifier" varchar(32) NOT NULL,"title" varchar(255) DEFAULT \'\' NOT NULL,CONSTRAINT "more_fields_as_pk_pkey" PRIMARY KEY ("identifier","subidentifier"));'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
         
     public function testIndexes()
@@ -134,7 +134,7 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index(['title', 'alias'], 'table_with_indexes_title_alias', 'unique')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index('bodytext', 'table_with_indexes_bodytext', 'fulltext', 'hash')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE SEQUENCE "table_with_indexes_seq";',
             'CREATE TABLE "table_with_indexes" ("id" int4 DEFAULT nextval(\'table_with_indexes_seq\'::regclass) NOT NULL,"title" varchar(255) NOT NULL,"alias" varchar(255) NOT NULL,"sorting" int4 NOT NULL,"bodytext" text NOT NULL,CONSTRAINT "table_with_indexes_pkey" PRIMARY KEY ("id"));',
@@ -142,7 +142,7 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
             'CREATE UNIQUE INDEX "table_with_indexes_title_alias" ON "table_with_indexes" ("title","alias");',
             'CREATE FULLTEXT INDEX "table_with_indexes_bodytext" ON "table_with_indexes" USING HASH ("bodytext");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testForeignKeys()
@@ -154,12 +154,12 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('foreign_table_id', 'integer')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addForeignKey(new ForeignKey('foreign_table_id', 'second_table')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE SEQUENCE "table_with_foreign_keys_seq";',
             'CREATE TABLE "table_with_foreign_keys" ("id" int4 DEFAULT nextval(\'table_with_foreign_keys_seq\'::regclass) NOT NULL,"title" varchar(255) NOT NULL,"alias" varchar(255) NOT NULL,"foreign_table_id" int4 NOT NULL,CONSTRAINT "table_with_foreign_keys_pkey" PRIMARY KEY ("id"),CONSTRAINT "table_with_foreign_keys_foreign_table_id" FOREIGN KEY ("foreign_table_id") REFERENCES "second_table" ("id"));'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testIndexesAndForeignKeys()
@@ -176,7 +176,7 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index(['title', 'alias'], 'table_with_indexes_and_foreign_keys_title_alias', 'unique')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index('bodytext', 'table_with_indexes_and_foreign_keys_bodytext', 'fulltext', 'hash')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE SEQUENCE "table_with_indexes_and_foreign_keys_seq";',
             'CREATE TABLE "table_with_indexes_and_foreign_keys" ("id" int4 DEFAULT nextval(\'table_with_indexes_and_foreign_keys_seq\'::regclass) NOT NULL,"title" varchar(255) NOT NULL,"alias" varchar(255) NOT NULL,"sorting" int4 NOT NULL,"bodytext" text NOT NULL,"foreign_table_id" int4 NOT NULL,CONSTRAINT "table_with_indexes_and_foreign_keys_pkey" PRIMARY KEY ("id"),CONSTRAINT "table_with_indexes_and_foreign_keys_foreign_table_id" FOREIGN KEY ("foreign_table_id") REFERENCES "second_table" ("foreign_id") ON DELETE SET NULL ON UPDATE SET NULL);',
@@ -184,18 +184,18 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
             'CREATE UNIQUE INDEX "table_with_indexes_and_foreign_keys_title_alias" ON "table_with_indexes_and_foreign_keys" ("title","alias");',
             'CREATE FULLTEXT INDEX "table_with_indexes_and_foreign_keys_bodytext" ON "table_with_indexes_and_foreign_keys" USING HASH ("bodytext");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->createTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
     }
     
     public function testDropTable()
     {
         $table = new Table('drop');
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'DROP TABLE "drop"',
             'DROP SEQUENCE IF EXISTS "drop_seq"',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->dropTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->dropTable($table));
     }
     
     public function testAlterTable()
@@ -205,33 +205,45 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('title', 'string')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('alias', 'string')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'ALTER TABLE "add_columns" ADD COLUMN "title" varchar(255) NOT NULL,ADD COLUMN "alias" varchar(255) NOT NULL;'
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->alterTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
+        
+        // add and remove primary key
+        $table = new Table('change_primary_key');
+        $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->dropPrimaryKey());
+        $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addPrimary('new_primary'));
+        
+        $queryBuilder = new PgsqlQueryBuilder();
+        $expectedQueries = [
+            'ALTER TABLE "change_primary_key" DROP CONSTRAINT "change_primary_key_pkey";',
+            'ALTER TABLE "change_primary_key" ADD CONSTRAINT "change_primary_key_pkey" PRIMARY KEY ("new_primary");',
+        ];
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
         
         // add index
         $table = new Table('add_index');
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index('alias', 'add_index_alias', 'unique')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'CREATE UNIQUE INDEX "add_index_alias" ON "add_index" ("alias");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->alterTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
         
         // add column and index
         $table = new Table('add_column_and_index');
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addColumn(new Column('alias', 'string')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index('alias', 'add_column_and_index_alias', 'unique')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'ALTER TABLE "add_column_and_index" ADD COLUMN "alias" varchar(255) NOT NULL;',
             'CREATE UNIQUE INDEX "add_column_and_index_alias" ON "add_column_and_index" ("alias");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->alterTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
         
         // add foreign key, index, columns
         $table = new Table('add_columns_index_foreign_key');
@@ -240,13 +252,13 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addIndex(new Index('sorting', 'add_columns_index_foreign_key_sorting')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addForeignKey(new ForeignKey('foreign_key_id', 'referenced_table')));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'ALTER TABLE "add_columns_index_foreign_key" ADD COLUMN "foreign_key_id" int4 NOT NULL,ADD COLUMN "sorting" int4 NOT NULL;',
             'CREATE INDEX "add_columns_index_foreign_key_sorting" ON "add_columns_index_foreign_key" ("sorting");',
             'ALTER TABLE "add_columns_index_foreign_key" ADD CONSTRAINT "add_columns_index_foreign_key_foreign_key_id" FOREIGN KEY ("foreign_key_id") REFERENCES "referenced_table" ("id");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->alterTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
         
         // remove columns
         
@@ -264,7 +276,7 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addForeignKey(new ForeignKey('foreign_key_id', 'referenced_table')));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->dropForeignKey('foreign_key_to_drop_id'));
         
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'DROP INDEX "all_in_one_alias";',
             'ALTER TABLE "all_in_one" DROP CONSTRAINT "all_in_one_foreign_key_to_drop_id";',
@@ -273,7 +285,7 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
             'CREATE INDEX "all_in_one_sorting" ON "all_in_one" ("sorting");',
             'ALTER TABLE "all_in_one" ADD CONSTRAINT "all_in_one_foreign_key_id" FOREIGN KEY ("foreign_key_id") REFERENCES "referenced_table" ("id");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->alterTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
         
         // mixed order of calls add / remove column, add / remove index, add / remove foreign key - output is the same
         $table = new Table('all_in_one_mixed');
@@ -285,7 +297,7 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->dropIndex('all_in_one_mixed_alias'));
         $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->addForeignKey(new ForeignKey('foreign_key_id', 'referenced_table')));
 
-        $queryCreator = new PgsqlQueryBuilder();
+        $queryBuilder = new PgsqlQueryBuilder();
         $expectedQueries = [
             'DROP INDEX "all_in_one_mixed_alias";',
             'ALTER TABLE "all_in_one_mixed" DROP CONSTRAINT "all_in_one_mixed_foreign_key_to_drop_id";',
@@ -294,6 +306,31 @@ class PgsqlQueryBuilderTest extends PHPUnit_Framework_TestCase
             'CREATE INDEX "all_in_one_mixed_sorting" ON "all_in_one_mixed" ("sorting");',
             'ALTER TABLE "all_in_one_mixed" ADD CONSTRAINT "all_in_one_mixed_foreign_key_id" FOREIGN KEY ("foreign_key_id") REFERENCES "referenced_table" ("id");',
         ];
-        $this->assertEquals($expectedQueries, $queryCreator->alterTable($table));
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
+    }
+
+    public function testChangeColumn()
+    {
+        $table = new Table('with_columns_to_change');
+        $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->changeColumn('old_name', new Column('new_name', 'integer')));
+        $this->assertInstanceOf('\Phoenix\Database\Element\Table', $table->changeColumn('no_name_change', new Column('no_name_change', 'integer')));
+        
+        $queryBuilder = new PgsqlQueryBuilder();
+        $expectedQueries = [
+            'ALTER TABLE "with_columns_to_change" RENAME COLUMN "old_name" TO "new_name";',
+            'ALTER TABLE "with_columns_to_change" ALTER COLUMN "new_name" TYPE int4 USING new_name::integer;',
+            'ALTER TABLE "with_columns_to_change" ALTER COLUMN "no_name_change" TYPE int4 USING no_name_change::integer;',
+        ];
+        $this->assertEquals($expectedQueries, $queryBuilder->alterTable($table));
+    }
+
+    public function testRenameTable()
+    {
+        $table = new Table('old_table_name');
+        $queryBuilder = new PgsqlQueryBuilder();
+        $expectedQueries = [
+            'ALTER TABLE "old_table_name" RENAME TO "new_table_name";',
+        ];
+        $this->assertEquals($expectedQueries, $queryBuilder->renameTable($table, 'new_table_name'));
     }
 }